Move dyn event ops to new module
